Move ObjectMapperProducer to runtime package as it's not API
